refactor(test): consolidate layout bench/golden generators into one helper
makeRng/mkFile/mkDir/genDir/genCommits/bboxOf and the digest hasher were
copy-pasted (with drift) across 5 bench + test files. Extract them to
tests/_helpers/layoutTreeFixtures.ts as canonical helpers — two named tree
generators (genWeightedTree, genNestedTree), flatTree, genCommits, bboxOf, and
makeDigestHasher. The golden digests are unchanged (the shared helpers are
functionally identical to each golden's originals); the profilers + equivalence
test align to the canonical generators. ~55% fewer lines across the 5 files.

Refs #63
