+++ app/bench/layoutProfile.bench.test.ts
@@ -11,74 +11,7 @@ import { describe, it } from 'vitest';
 import { layoutCity } from '@/city/layout/algorithm.js';
 import { setLayoutProfiling, getLayoutProfile } from '@/city/layout/profiling';
 import { NodeKind } from '@/types';
-
-// Deterministic LCG so tree shape (and thus timings) are reproducible run-to-run.
-function makeRng(seed: number): () => number {
-  let s = seed >>> 0;
-  return () => {
-    s = (s * 1664525 + 1013904223) >>> 0;
-    return s / 0x100000000;
-  };
-}
-
-function mkFile(name: string, path: string, rng: () => number): any {
-  return {
-    name,
-    type: NodeKind.File,
-    path: `${path}/${name}`,
-    extension: '.c',
-    size: 200 + Math.floor(rng() * 40000),
-    lines: 5 + Math.floor(rng() * 1500),
-    created: '2024-01-01T00:00:00Z',
-    modified: '2024-01-01T00:00:00Z',
-  };
-}
-
-function mkDir(name: string, children: any[], path: string): any {
-  const descendants_count =
-    children.length + children.reduce((acc, c) => acc + (c.descendants_count || 0), 0);
-  return {
-    name,
-    type: NodeKind.Directory,
-    path,
-    children_count: children.length,
-    descendants_count,
-    descendants_size: 1000,
-    children,
-  };
-}
-
-// genDir(name, path, budget, depth, rng) — recursively splits a file budget
-// across loose files + skewed subdirectories, mimicking a real source tree.
-function genDir(name: string, path: string, budget: number, depth: number, rng: () => number): any {
-  if (budget <= 8 || depth >= 7) {
-    const files = [];
-    for (let i = 0; i < budget; i++) files.push(mkFile(`f${i}.c`, path, rng));
-    return mkDir(name, files, path);
-  }
-  const children: any[] = [];
-  const loose = 1 + Math.floor(rng() * 6);
-  for (let i = 0; i < loose; i++) children.push(mkFile(`a${i}.c`, path, rng));
-  let remaining = budget - loose;
-  const nSub = 2 + Math.floor(rng() * 6);
-  // Power-law weights → first child gets the lion's share (drivers/-like skew).
-  const weights: number[] = [];
-  let wsum = 0;
-  for (let i = 0; i < nSub; i++) {
-    const w = 1 / (i + 1);
-    weights.push(w);
-    wsum += w;
-  }
-  for (let i = 0; i < nSub && remaining > 0; i++) {
-    const last = i === nSub - 1;
-    let share = last ? remaining : Math.max(1, Math.round((weights[i] / wsum) * (budget - loose)));
-    share = Math.min(share, remaining);
-    remaining -= share;
-    const cname = `d${i}`;
-    children.push(genDir(cname, `${path}/${cname}`, share, depth + 1, rng));
-  }
-  return mkDir(name, children, path);
-}
+import { makeRng, genWeightedTree } from '../tests/_helpers/layoutTreeFixtures';
 
 function countFilesDirs(node: any): { files: number; dirs: number } {
   let files = 0;
@@ -97,7 +30,7 @@ function countFilesDirs(node: any): { files: number; dirs: number } {
 describe('layoutCity profile (Linux-sized)', () => {
   function runOne(label: string, budget: number) {
     const rng = makeRng(0xc0ffee);
-    const tree = genDir('root', 'root', budget, 0, rng);
+    const tree = genWeightedTree('root', 'root', budget, 0, rng);
     const { files, dirs } = countFilesDirs(tree);
     setLayoutProfiling(true);
     const t0 = performance.now();
